Refactor pipelines context and add useSafePipelines hook
